refactor(KB-264): extract pipeline tracking helpers to lib/pipeline-tracking.js
- Move ensurePipelineRun, startStepRun, completeStepRun, failStepRun, skipStepRun to separate module
- agent-jobs.js now 413 lines (was 533), under 500 line limit
- Improves code organization and reusability
